fix: move ReactSliderShadowNode to oldarch
